Run a mixed benchmark workload
Originally we have used `rand.Int()` to determine whether to enqueue or
dequeue and assumed that it would return both negative and positive
numbers whereas in reality it returns only non-negative numbers. With
this commit we use `rand.IntN(n int)` instead to control the range and
enqueue/dequeue with a 50% chance. We also avoid overcommitting Go
routines and only use the number of available CPUs as this benchmark is
CPU bound anyway. The number of assigned CPUs should be controlled via
the parameter `-cpu` on the command line.